pyproject.toml
[project]
name = "diagram-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "diagrams>=0.24.4",
    "fastapi>=0.116.1",
    "gradio>=5.0.0",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "requests>=2.32.4",
]
